separated statistics extraction and statistics plotting into two scripts
